third_party/WebKit/LayoutTests/inspector/console/console-format-style-expected.txt

CONSOLE MESSAGE: line 8: %cBlue!.
CONSOLE MESSAGE: line 9: %cBlue! %cRed!
Tests that console logging dumps properly styled messages.

console-format-style.html:8 Blue!.
Styled text #0: color: blue;
console-format-style.html:9 Blue! Red!
Styled text #0: color: blue;
Styled text #1: color: red;
